requirements.txt
python-telegram-bot==21.6
openai
gspread
google-auth
pytz
python-dotenv
notion-client
